Fix use of modals for host info error
When host addition failed, the error was presented
as an JSON object containing jsx - but this caused
problem when the user clicked delete to remove the
problematic row. The fix introduces a new class
for the host info element, which builds the jsx
correctly. This also allows for further modal
based help with other error in the future.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>