Table layout fixed
Probe table did use fixed thead to handle the
scrolling of the tbody sections. This patch drops
this approach in favor of two tables with the 2nd
table being the scrollable component.

In addition, now when yo navigate back to a prior
page the inline-notification, and probe button
state is reset.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>