Add support for iSCSI networking
With iscsi being added, there are now 4 network
subnets potentially within the cluster so instead
making the cluster table longer, all the network
options now sit in there own widget.

In addition the host listing table has been
extended to include the IP of the iscsi target
that will be used on the host.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>