podman-docker:
With this modification an error is printed if there is no container manager installed.

Also this modification cover the case when podman-docker.noarch is installed.
In this situation, the script was cheated, and the container manager selected was <docker> instead <podman>,
making imposible to start the ARS container because wrong options selected.
Simply changing the order in the <CONTAINER_OPTS> var solves this issue.