Remove prometheus port handling logic
In older versions of ceph-ansible the default
prometheus_port was 9090, which conflicts with
cockpit's default port. This meant checking for
this conflist and adjusting the prometheus_port
setting. ceph-ansible has now changed in v4 to
use port 9092, so the conflict is gone, and the
code in the UI can be removed.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>