Accomodate 4k sectors when calculating capacity
Code was using sector*sectorsize to determine the
capacity, but on some kernels ansible is unable
to determine this correctly. To avoid the issue,
size is now reverse engineered from the 'size'
attribute of a device (str -> bytes).

In addition, capacities over 1TB now use 1 decimal
place to make them more meaningful

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>