Saves runtime settings at deploy time
All settings chosen in the UI are saved in a file
in the cockput users home directory to provide
a record of the environment at install time. The
data stored also includes the probe results for
each host, so any warnings/errors detected by
the probe process are also recorded.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>