utils: pull image from registry.redhat.io
Prior to this change, we would pull the image from whatever registries
the user defined in their local docker configuration.

On RHEL 7, that could mean that we pull the image from a registry other
than registry.redhat.io.

Specify the registry directly so that we pull the product container from
the correct location.