Improve ssh key handling for sudo users
This patch improve sudo users ssh key handling as
reported in BZ1791143. Now if the users ssh key (RSA)
only exists, it will be applied to the runner-service, and
if the sudo user doesn't have keys set up, the process
will seed ssh priv/pub keys from the runner-service to
the user.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>